test: add reusable mock db helper
