+++ package.json
@@ -20,5 +20,8 @@
     "jest": "^30.0.5",
     "jest-environment-jsdom": "^30.0.5",
     "prettier": "^3.6.2"
+  },
+  "jest": {
+    "testMatch": ["**/?(*.)+(spec|test).[jt]s?(x)"]
   }
 }
